v0.21
BranchPage and api endpoint to get list of files for given branch in repository created. Also api for creating files is integrated in frontend to upload files to given branch of a repository. DB table structure changed so that storing and comparing files is similar in functionality to that of github. Now using hash of file contents to avoid storing redundant data even if the version if different.