v0.25
Confirmation dialogue state shifted from redux store to react context as it contained a JSX component in one of its state which isn't used in redux as it is not serializable. A new file called RepositoryCode created from RepostoryPage to handle file and folder viewing in future versions to have ui like github to navigate between files in the project structure.